Copying image array elementwise
This is a first (non-efficient) implementation of the image
wrapper of the rust array. Added an C accessor for element values
to simplify pointer arithmetic (although pointer arithmetic is not
natively supported in Go, it is possible to do it through
unsafe.Pointer; however the syntax becomes terribly verbose).

Image is still missaligned. It is still necessary to figure out
the ownership of the underlying memory once rust returns the const
pointer (?).