Rendering from wasm
Access to the actual wasm memory buffer is unclear with the
current wasm-pack output. It was necessary to manually export
`export let wasm_memory = wasm.memory` in rendering_wasm.js.
The buffer (and the pointer to the starting point of the frame
in that buffer) is then mapped to a Uint8ClampedArray and an
ImageData object.

wasm-pack should be configured to automatically expose the
modules memory buffer.

Several good references:

* https://www.hellorust.com/demos/canvas/index.html
* https://rustwasm.github.io/docs/wasm-bindgen/print.html
* https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/putImageData