Added random scene (book cover).
Fixed issue with lens blur which would only show up in the new scene
(given that the aperture was approaching zero). My random sampling was
relying on the translation embedded in the eye_to_world transformation
however, as it approached zero (due to the aperture), both the direction
and origin of the ray were also. To fix it, the origin and offset are
split such that only the random offset tends to zero with the aperture.